fix: completely separated APIs ans Socket server
