+++ .env-example
@@ -8,5 +8,6 @@ AUTH_KEY=auth.key
 SESSION_SECRET=session.secret
 EMAIL_ADDRESS=email.address
 EMAIL_PASSWORD=email.password
-
+SOCKET_PORT=socket.io.port
+HOST=host
