+++ package.json
@@ -6,7 +6,9 @@
   "scripts": {
     "build": "npm run lint:fix && tsc",
     "start": "node build/src/index.js",
-    "dev": "nodemon",
+    "start:socket": "node build/src/socket.js",
+    "dev": "nodemon --files src/index.ts",
+    "dev:socket": "nodemon --files src/socket.ts",
     "test": "cross-env NODE_ENV=test nyc mocha --require ts-node/register --recursive test/**/*.spec.ts --timeout 20000 --exit",
     "lint": "eslint .",
     "lint:fix": "eslint --fix ."
