+++ src/socket/Dockerfile
@@ -1,0 +1,13 @@
+FROM node:16-alpine3.16
+
+WORKDIR /usr/src/app
+
+COPY . .
+
+RUN npm install
+
+RUN npm run build
+
+EXPOSE 4003
+
+CMD [ "npm", "start:socket" ]
