+++ types/index.ts
@@ -1,1 +1,2 @@
 export * from "./express"
+export * from "./socket"
